[#8] Frontend module (+Space fix)
Mainly "fixing" Gael's work because of a weird visual issue on my end.
Worked around gap-3-4-5 and mt-4, as the space between the divs don't seem to work properly here, maybe that it's tied to how the symfony forms work ?
Needs to be looked into if we were to make other frontend changes.